훅 분리 2
