+++ src/page/Product.tsx
@@ -1,6 +1,6 @@
 import { DefaultDiv } from "@/styles/Common.styled"
 import BottomBar from "@/component/product/BottomBar";
-import ProductThumbNail from "@/component/product/ProductThumbNail";
+import ProductThumbNail from "@/component/product/ProductThumbnail ";
 import ProductDetail from "@/component/product/ProductDetail";
 import { ErrorBoundary } from "react-error-boundary";
 import { Suspense } from "react";
